reorganized code structure
